+++ changelog.md
@@ -3,6 +3,9 @@
 
 ## JHelioviewer 4.7.4 (pending)
 
+- Make strength of radial enhancement configurable
+- Add ability for direct setting of FITS data clipping range
+
 ## JHelioviewer 4.7.3 (2025-07-01)
 
 - Improve performance on ARM Macs